# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders": {
      "post": {
        "operationId": "OrderController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully"
          },
          "400": {
            "description": "Invalid order data or coupon validation failed"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create order with applied coupon",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "delete": {
        "operationId": "OrderController_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled successfully"
          },
          "400": {
            "description": "Order not found or already cancelled"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Cancel an order",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coupon Management System API",
    "description": "API documentation for Coupon Code Management System - Backend Technical Challenge",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "prod-123",
            "description": "Product ID"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity of items",
            "minimum": 1
          },
          "price": {
            "type": "number",
            "example": 500,
            "description": "Price per item",
            "minimum": 0
          },
          "category": {
            "type": "string",
            "example": "electronics",
            "description": "Product category"
          }
        },
        "required": [
          "productId",
          "quantity",
          "price"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "orderValue": {
            "type": "number",
            "example": 1000,
            "description": "Total order value before discount",
            "minimum": 0
          },
          "items": {
            "description": "Order items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "paymentMethod": {
            "type": "string",
            "example": "card",
            "description": "Payment method (card, upi, wallet)"
          },
          "couponCode": {
            "type": "string",
            "example": "FIRST50",
            "description": "Optional coupon code to apply"
          },
          "deliveryCharge": {
            "type": "number",
            "example": 50,
            "description": "Delivery charge",
            "minimum": 0
          }
        },
        "required": [
          "orderValue",
          "items",
          "paymentMethod"
        ]
      },
...
    }
  }
}